add hadoop

+++ hadoop/src/main/java/com/huma/hadoop/reducer/WordCountReducer.java
@@ -1,5 +1,6 @@
 package com.huma.hadoop.reducer;
 
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
@@ -10,13 +11,13 @@ import java.io.IOException;
  * @author hudenian
  * @date 2021/6/24
  */
-public class WordCountReducer extends Reducer<Text, LongWritable, Text, LongWritable> {
-    private LongWritable result = new LongWritable();
+public class WordCountReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
+    private IntWritable result = new IntWritable();
 
     @Override
-    public void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException {
+    public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
         int sum = 0;
-        for (LongWritable val : values) {
+        for (IntWritable val : values) {
             sum += val.get();
         }
         this.result.set(sum);
